refactor: Extract csv printer factory
