+++ dflib-csv/src/test/java/org/dflib/csv/CsvSaverTest.java
@@ -3,6 +3,8 @@ package org.dflib.csv;
 import org.dflib.DataFrame;
 import org.dflib.csv.parser.format.CsvFormat;
 import org.dflib.csv.parser.format.LineBreak;
+import org.dflib.csv.printer.DefaultCsvPrinter;
+import org.dflib.csv.printer.CsvPrinterFactory;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -23,5 +25,12 @@ public class CsvSaverTest {
     @Test
     public void saveToString_NoHeader() {
         assertEquals("1,2\n3,4\n", Csv.saver().format(FORMAT).noHeader().saveToString(DF));
+    }
+
+    @Test
+    public void saveToString_CustomPrinterFactory() {
+        CsvPrinterFactory factory = config -> new DefaultCsvPrinter(config, (value, out) -> out.append('x'));
+
+        assertEquals("x,x\nx,x\nx,x\n", Csv.saver().printerFactory(factory).format(FORMAT).saveToString(DF));
     }
 }
